feat: Unary push-down for symmetric float literals in forge emitters
- Add Unary{Neg|Pos} push-down to 5 emitters so -40 in float context becomes -40.0 instead of -40 as f64 or (-40).toDouble()
- Mirror existing Binary push-down by recursing into operand with the float expected
- 159 unchanged byte-goldens prove the push-down is scoped to Unary{Neg, NumberLit} only

+++ tests/forge/expected/CrossfileValidatorTransform.kt
@@ -13,7 +13,7 @@ class CrossfileValidatorTransform {
     fun validate(rawTemp: UShort): ValidationResult {
         if (rawTemp.toInt() < 0 || rawTemp.toInt() > 4095)
             return ValidationResult(false, "raw_temp_out_of_range")
-        if (!(computeTemperature(rawTemp) > (-40).toDouble() && computeTemperature(rawTemp) < 200.0))
+        if (!(computeTemperature(rawTemp) > -40.0 && computeTemperature(rawTemp) < 200.0))
             return ValidationResult(false, "plausibility_failed")
         return ValidationResult(true, "")
     }
